backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
Pillow
django-cors-headers